administrator-roles

+++ Modules/AdministratorModule/resources/views/roles/table_action.blade.php
@@ -1,12 +1,16 @@
 <div class="form-button-action">
-    <a class="btn btn-link btn-primary btn-lg" href="{{ route('administrator-roles.edit', $row->id) }}">
-        <i class="fa fa-edit"></i>
-    </a>
-    <button type="button" class="btn btn-link btn-danger"
-        onclick="destroyAction(
-        '{{ route('administrator-roles.destroy', $row->id) }}', 
-        'Peran {{ $row->name }}'
-    )">
-        <i class="fa fa-times"></i>
-    </button>
+    @can('administrator-roles.edit')
+        <a class="btn btn-link btn-primary btn-lg" href="{{ route('administrator-roles.edit', $row->id) }}">
+            <i class="fa fa-edit"></i>
+        </a>
+    @endcan
+    @can('administrator-roles.destroy')
+        <button type="button" class="btn btn-link btn-danger"
+            onclick="destroyAction(
+            '{{ route('administrator-roles.destroy', $row->id) }}', 
+            'Peran {{ $row->name }}'
+        )">
+            <i class="fa fa-times"></i>
+        </button>
+    @endcan
 </div>
